+++ compare258.c
@@ -9,8 +9,8 @@
 #include "fallback_builtins.h"
 
 /* ALIGNED, byte comparison */
-static inline int32_t compare256_c_static(const unsigned char *src0, const unsigned char *src1) {
-    int32_t len = 0;
+static inline uint32_t compare256_c_static(const unsigned char *src0, const unsigned char *src1) {
+    uint32_t len = 0;
 
     do {
         if (*src0 != *src1)
@@ -42,7 +42,7 @@ static inline int32_t compare256_c_static(const unsigned char *src0, const unsig
     return 256;
 }
 
-static inline int32_t compare258_c_static(const unsigned char *src0, const unsigned char *src1) {
+static inline uint32_t compare258_c_static(const unsigned char *src0, const unsigned char *src1) {
     if (*src0 != *src1)
         return 0;
     src0 += 1, src1 += 1;
@@ -53,7 +53,7 @@ static inline int32_t compare258_c_static(const unsigned char *src0, const unsig
     return compare256_c_static(src0, src1) + 2;
 }
 
-Z_INTERNAL int32_t compare258_c(const unsigned char *src0, const unsigned char *src1) {
+Z_INTERNAL uint32_t compare258_c(const unsigned char *src0, const unsigned char *src1) {
     return compare258_c_static(src0, src1);
 }
 
@@ -65,8 +65,8 @@ Z_INTERNAL int32_t compare258_c(const unsigned char *src0, const unsigned char *
 
 #ifdef UNALIGNED_OK
 /* UNALIGNED_OK, 16-bit integer comparison */
-static inline int32_t compare256_unaligned_16_static(const unsigned char *src0, const unsigned char *src1) {
-    int32_t len = 0;
+static inline uint32_t compare256_unaligned_16_static(const unsigned char *src0, const unsigned char *src1) {
+    uint32_t len = 0;
 
     do {
         if (*(uint16_t *)src0 != *(uint16_t *)src1)
@@ -86,14 +86,14 @@ static inline int32_t compare256_unaligned_16_static(const unsigned char *src0, 
     return 256;
 }
 
-static inline int32_t compare258_unaligned_16_static(const unsigned char *src0, const unsigned char *src1) {
+static inline uint32_t compare258_unaligned_16_static(const unsigned char *src0, const unsigned char *src1) {
     if (*(uint16_t *)src0 != *(uint16_t *)src1)
         return (*src0 == *src1);
 
     return compare256_unaligned_16_static(src0+2, src1+2) + 2;
 }
 
-Z_INTERNAL int32_t compare258_unaligned_16(const unsigned char *src0, const unsigned char *src1) {
+Z_INTERNAL uint32_t compare258_unaligned_16(const unsigned char *src0, const unsigned char *src1) {
     return compare258_unaligned_16_static(src0, src1);
 }
 
@@ -105,8 +105,8 @@ Z_INTERNAL int32_t compare258_unaligned_16(const unsigned char *src0, const unsi
 
 #ifdef HAVE_BUILTIN_CTZ
 /* UNALIGNED_OK, 32-bit integer comparison */
-static inline int32_t compare256_unaligned_32_static(const unsigned char *src0, const unsigned char *src1) {
-    int32_t len = 0;
+static inline uint32_t compare256_unaligned_32_static(const unsigned char *src0, const unsigned char *src1) {
+    uint32_t len = 0;
 
     do {
         uint32_t sv = *(uint32_t *)src0;
@@ -115,7 +115,7 @@ static inline int32_t compare256_unaligned_32_static(const unsigned char *src0, 
 
         if (diff) {
             uint32_t match_byte = __builtin_ctz(diff) / 8;
-            return (int32_t)(len + match_byte);
+            return len + match_byte;
         }
 
         src0 += 4, src1 += 4, len += 4;
@@ -124,14 +124,14 @@ static inline int32_t compare256_unaligned_32_static(const unsigned char *src0, 
     return 256;
 }
 
-static inline int32_t compare258_unaligned_32_static(const unsigned char *src0, const unsigned char *src1) {
+static inline uint32_t compare258_unaligned_32_static(const unsigned char *src0, const unsigned char *src1) {
     if (*(uint16_t *)src0 != *(uint16_t *)src1)
         return (*src0 == *src1);
 
     return compare256_unaligned_32_static(src0+2, src1+2) + 2;
 }
 
-Z_INTERNAL int32_t compare258_unaligned_32(const unsigned char *src0, const unsigned char *src1) {
+Z_INTERNAL uint32_t compare258_unaligned_32(const unsigned char *src0, const unsigned char *src1) {
     return compare258_unaligned_32_static(src0, src1);
 }
 
@@ -145,8 +145,8 @@ Z_INTERNAL int32_t compare258_unaligned_32(const unsigned char *src0, const unsi
 
 #if defined(UNALIGNED64_OK) && defined(HAVE_BUILTIN_CTZLL)
 /* UNALIGNED64_OK, 64-bit integer comparison */
-static inline int32_t compare256_unaligned_64_static(const unsigned char *src0, const unsigned char *src1) {
-    int32_t len = 0;
+static inline uint32_t compare256_unaligned_64_static(const unsigned char *src0, const unsigned char *src1) {
+    uint32_t len = 0;
 
     do {
         uint64_t sv = *(uint64_t *)src0;
@@ -155,7 +155,7 @@ static inline int32_t compare256_unaligned_64_static(const unsigned char *src0, 
 
         if (diff) {
             uint64_t match_byte = __builtin_ctzll(diff) / 8;
-            return (int32_t)(len + match_byte);
+            return len + match_byte;
         }
 
         src0 += 8, src1 += 8, len += 8;
@@ -164,14 +164,14 @@ static inline int32_t compare256_unaligned_64_static(const unsigned char *src0, 
     return 256;
 }
 
-static inline int32_t compare258_unaligned_64_static(const unsigned char *src0, const unsigned char *src1) {
+static inline uint32_t compare258_unaligned_64_static(const unsigned char *src0, const unsigned char *src1) {
     if (*(uint16_t *)src0 != *(uint16_t *)src1)
         return (*src0 == *src1);
 
     return compare256_unaligned_64_static(src0+2, src1+2) + 2;
 }
 
-Z_INTERNAL int32_t compare258_unaligned_64(const unsigned char *src0, const unsigned char *src1) {
+Z_INTERNAL uint32_t compare258_unaligned_64(const unsigned char *src0, const unsigned char *src1) {
     return compare258_unaligned_64_static(src0, src1);
 }
